Add files via upload
Download files to a folder. Open terminal in the folder with the files. Type "make" and then press enter. Then type "./a.out cameron.ged" to execute and read the .ged genealogy file. It will create a file "sorted.dat" that has sorted the names alphabetically and then display the family line of the the first person in the file on the command line.